split inline_anchors
