+++ lib/asciidoctor/rfc/v3/base.rb
@@ -128,71 +128,6 @@ module Asciidoctor
         end
       end
 
-      def inline_anchor(node)
-        case node.type
-        when :xref
-          text = node.text
-          if text =~ /^\S+ (of|comma|parens|bare)\b/
-            # <<crossreference#fragment,section (of|comma|parens|bare): text>> = relref
-            matched = /(?<section>\S+)\s+(?<format>[a-z]+)(: )?(?<text>.*)$/.match node.text
-
-            relref_contents = matched[:text]
-
-            relref_attributes = {
-              relative: node.attributes["fragment"],
-              section: matched[:section],
-              displayFormat: matched[:format],
-              target: node.target.gsub(/\..*$/, "").gsub(/^#/, ""),
-            }.reject { |_, value| value.nil? }
-
-            noko do |xml|
-              xml.relref relref_contents, **relref_attributes
-            end.join
-          else
-            matched = /^format=(?<format>counter|title|none|default):\s*(?<text>.*)$/.match node.text
-            matched ||= {}
-            xref_contents = matched[:text] || node.text
-
-            xref_attributes = {
-              format: matched[:format],
-              target: node.target.gsub(/^#/, ""),
-            }.reject { |_, value| value.nil? }
-
-            noko do |xml|
-              xml.xref xref_contents, **xref_attributes
-            end.join
-          end
-        when :link
-          eref_contents = node.target == node.text ? nil : node.text
-
-          eref_attributes = {
-            target: node.target,
-          }.reject { |_, value| value.nil? }
-
-          noko do |xml|
-            xml.eref eref_contents, **eref_attributes
-          end.join
-        when :bibref
-          unless node.xreftext.nil?
-            x = node.xreftext.gsub(/^\[(.+)\]$/, "\\1")
-            if node.id != x
-              $xreftext[node.id] = x
-            end
-          end
-          # NOTE technically node.text should be node.reftext, but subs have already been applied to text
-          %(<bibanchor="#{node.id}">) # will convert to anchor attribute upstream
-        when :ref
-          # If this is within referencegroup, output as bibanchor anyway
-          if $processing_reflist
-            %(<bibanchor="#{node.id}">) # will convert to anchor attribute upstream
-          else
-            warn %(asciidoctor: WARNING: anchor "#{node.id}" is not in a place where XML RFC will recognise it as an anchor attribute)
-          end
-        else
-          warn %(asciidoctor: WARNING: unknown anchor type: #{node.type.inspect})
-        end
-      end
-
       # Syntax:
       #   [[id]]
       #   [keepWithNext=true,keepWithPrevious=true] (optional)
