+++ lib/asciidoctor/rfc/v3/converter.rb
@@ -4,9 +4,10 @@ require "asciidoctor/rfc/version"
 require "asciidoctor/rfc/common/base"
 require "asciidoctor/rfc/common/front"
 require "asciidoctor/rfc/v3/base"
-require "asciidoctor/rfc/v3/front"
-require "asciidoctor/rfc/v3/lists"
 require "asciidoctor/rfc/v3/blocks"
+require "asciidoctor/rfc/v3/front"
+require "asciidoctor/rfc/v3/inline_anchor"
+require "asciidoctor/rfc/v3/lists"
 require "asciidoctor/rfc/v3/table"
 
 module Asciidoctor
@@ -23,9 +24,10 @@ module Asciidoctor
       include ::Asciidoctor::RFC::Common::Base
       include ::Asciidoctor::RFC::Common::Front
       include ::Asciidoctor::RFC::V3::Base
-      include ::Asciidoctor::RFC::V3::Front
-      include ::Asciidoctor::RFC::V3::Lists
       include ::Asciidoctor::RFC::V3::Blocks
+      include ::Asciidoctor::RFC::V3::Front
+      include ::Asciidoctor::RFC::V3::InlineAnchor
+      include ::Asciidoctor::RFC::V3::Lists
       include ::Asciidoctor::RFC::V3::Table
 
       register_for "rfc3"
